requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
python-dateutil>=2.8.2